feat: 직렬화 가능하도록 설정 변경
