+++ src/main/java/coupon/domain/Coupon.java
@@ -9,7 +9,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import java.io.Serializable;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.time.LocalDateTime;
@@ -18,7 +17,7 @@ import java.time.LocalTime;
 @NoArgsConstructor
 @Getter
 @Entity
-public class Coupon implements Serializable {
+public class Coupon {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
